requirements.txt
streamlit
google-genai
langchain
langchain-community
langchain-core
faiss-cpu
pypdf
python-dotenv
tiktoken
sentence-transformers
